Add `Place.description`.
Oh hey, it turns out that the `ImageSet.description` field is not
actually used in WWT! It appears in some of the files and docs, but the
engine and Windows client don't actually load it. The clients don't do
anything with `Place.description` either, but at least they load it.