Support glob arguments on Windows
On Windows, the basic command prompt doesn't expand globs. It turns out
that if you want to support globs on this OS, you have to ... just
manually implement them. Annoying! But not so hard to do.

Note that we can't implement this as a `type` keyword because it
operates item-by-item, so it can't conveniently turn one item into many.
I think.