Work around missing members in the API documentation
I had been feeling like our API docs were missing a bunch of stuff that
they were supposed to have, and it turns out that we're facing a bug in
sphinx-automodapi:

https://github.com/astropy/sphinx-automodapi/issues/132

Here we work around the issue by turning on `:inherited-members:`
everywhere. It pollutes our namespaces a bunch but is better than having
nothing.