wwt_data_formats/imageset.py: straighten out WCS and parity
Revisit the WCS processing to make it a lot better, especially regarding
image parity.

WCS data contain information about image parity, which gets expressed in
the sign of the determinant of the "CD" matrix. This should get
propagated into the "BottomsUp" setting for WWT imagesets, but we
weren't doing this properly. We now deal with this properly ... which is
a breaking change, because it will severely alter the processing of
broken WCSes!

In so doing we also fix up small-but-important things like the handling
of "PC"-form matrices and some off-by-one errors. I think that the code
is much more correct now.

We also become stricter about input validation, because in our more
matrix-y approach it becomes clear that some setups cannot be expressed
in WWT's more limited vocabulary. In particular, it looks like you're
not allowed to have a tiled imageset with BottomsUp=true, in the sense
that they just won't render. So make sure to detect and reject those.