wwt_data_formats/imageset.py: fix verification of rotations near 45 degrees
The existing code was failing on NOIRLab `noao-abell39hotelling`, with
the following inputs:

```
   Scale:
      * -0.000216021137021
      * 0.000215968768353
   Rotation: 134.06957586364
```

Because the CD matrix works out to:

```
>> cd1_1, cd2_2 = 0.00015024947791364, 0.00015021305386212
>> cd1_2, cd2_1 = 0.00015517263725712, -0.00015521026392131
```

The determinant of the CD matrix has units [CD]^2, so our check was in
units of [CD]^-1, exceeding the tolerance. I believe that the correct
fix is to compare to the square root of the absolute determinant so that
we're dimensionally correct. To be honest I forget why I was normalizing
by the determinant here and not something like the RMS, but this new
version should be strictly better than what we had before.