Attempt to fix offset_y once and for all
We weren't getting it right for untiled images, still. I found a nice
sample image in the WWT corpus with a pretty large offset, so that
I could construct a test case that I could verify by looking at the
final on-sky rendering. I think we now do it all right.

Note that the sample image isn't actually used in the test suite, but
I'm adding it for manual visual verification, as documented extensively
in the new test case.