Make sure that Windows can load our imagesets
When loading the `imagesets.xml` file, the Windows app expects that
certain XML attributes and children are present. It will actually refuse
to start up if they're missing, complaining about a corrupted file.
Here we update our XML export to make sure that the <Imageset>s that
we emit meet the app's expectations. Technically this doesn't need to
happen for *all* imagesets that we emit, I think, but life will be
simplest if we just make sure that everything complies.

Note that even if we update the app to be more lenient, old versions of
the app are still going to be out there, so that's not a fully robust
solution.