wwt_data_formats/imageset.py: fix positioning of odd-sized studies
As explored in WorldWideTelescope/wwt-webgl-engine#212, there were times
when "study" FITS images were landing a half-pixel off from their
intended locations. After some digging, I realized that the issue is
that odd-sized images have to be shifted a half-pixel relative to their
"ideal" centering in the tiling pixelization. We need to take account of
that here.