wwt_data_formats/imageset.py: be less picky about WCS when we're TOAST
If we're using `set_position_from_wcs()` for a TOAST imageset, the
details of the WCS aren't very important, since we're just using it to
obtain basic centering and sizing info. So, we can skip various checks
about square pixels and such.

Also improve an error message that referred to a non-positive CD
determinant when it was really checking for 0 or NaN (in effect). We are
OK with negative determinants if the image isn't tiled.